add methods

+++ src/main/java/com/course/project/arina/dto/DeviceDTO.java
@@ -19,7 +19,7 @@ public class DeviceDTO {
     Status status;
     String name;
     String brand;
-    BigDecimal cost;
+    Double cost;
     LocalDate acceptanceDate;
     User user;
